+++ test/legacy_test/reader/CMakeLists.txt
@@ -1,1 +1,2 @@
-add_subdirectory(tests)
+py_test(decorator_test SRCS decorator_test.py)
+set_tests_properties(decorator_test PROPERTIES TIMEOUT 120)
