+++ test/legacy_test/reader/decorator_test.py
