--- test/legacy_test/reader/tests/CMakeLists.txt
@@ -1,2 +1,0 @@
-py_test(decorator_test SRCS decorator_test.py)
-set_tests_properties(decorator_test PROPERTIES TIMEOUT 120)
